Crate rerun_sdk

The Rerun SDK

Feature flags

re_viewer (enabled by default) — Support for the viewer

web — Support serving a web viewer over HTTP.

Enabling this adds quite a bit to the compile time and binary size, since it requires compiling and bundling the viewer as wasm.

Re-exports

pub use time::log_time;

Modules

external

global 🔒

msg_sender 🔒

session 🔒

time

viewer

Structs

AnnotationContext

The AnnotationContext provides additional information on how to display entities.

ApplicationId

The user-chosen name of the application doing the logging.

Arrow3D

A 3D Arrow

Box3D

A 3D bounding box represented by it’s half-lengths

ClassId

A 16-bit ID representing a type of semantic class.

ColorRGBA

An RGBA color tuple.

ComponentBundle

A ComponentBundle holds an Arrow component column, and its field name.

ComponentName

The name of an entity component, e.g. pos or color.

EncodedMesh3D

Compressed/encoded mesh format

EntityPath

camera / "left" / points / #42

InstanceKey

A number used to specify a specific instance in an entity.

KeypointId

A 16-bit ID representing a type of semantic keypoint within a class.

Label

A String label component

LineStrip2D

A Line Strip of 2D points

LineStrip3D

A Line Strip of 3D points

Mat3x3

A 3x3 column-major Matrix made up of 3 Vecs

MeshId

A unique id per Mesh3D.

MsgBundle

A MsgBundle holds data necessary for composing a single log message.

MsgId

A unique id per crate::LogMsg.

MsgSender

Facilitates building and sending component payloads with the Rerun SDK.

Pinhole

Camera perspective projection (a.k.a. intrinsics).

Point2D

A point in 2D space.

Point3D

A point in 3D space.

Quaternion

A Quaternion represented by 4 real numbers.

Radius

A Radius component

RawMesh3D

A raw “triangle soup” mesh.

RecordingId

A unique id per recording (a stream of LogMsges).

Rigid3

A proper rigid 3D transform, i.e. a rotation and a translation.

Scalar

A double-precision scalar.

ScalarPlotProps

Additional properties of a scalar when rendered as a plot.

Session

SignedAxis3

Size3D

Tensor

A Multi-dimensional Tensor

TensorDimension

Flattened Tensor data payload

TensorId

A unique id per Tensor.

TextEntry

A text entry component, comprised of a text body and its log level.

Time

A date-time represented as nanoseconds since unix epoch

TimeInt

Either nanoseconds or sequence numbers.

TimePoint

A point in time.

Timeline

A time frame/space, e.g. log_time or frame_nr, coupled with the type of time it keeps.

Vec2D

A vector in 2D space.

Vec3D

A vector in 3D space.

Vec4D

A vector in 4D space.

ViewCoordinates

How we interpret the coordinate system of an entity/space.

Enums

Axis3

Handedness

LogMsg

The most general log message sent from the SDK to the server.

Mesh3D

A Generic 3D Mesh.

MeshFormat

MsgSenderError

Rect2D

A rectangle in 2D space.

Sign

TensorData

Flattened Tensor data payload

TensorDataMeaning

TimeType

The type of a TimeInt.

Transform

A transform between two spaces.

Traits

Component

SerializableComponent

A trait to identify any Component that is ready to be collected and subsequently serialized into an Arrow payload.

TensorTrait

Functions

default_server_addr

global_session

Access a global Session singleton for convenient logging.